+++ test/nextjournal/clerk/viewer_test.clj
@@ -338,6 +338,30 @@
       (is (every? (every-pred not-empty string?) ids))
       (is (distinct? ids))))
 
+  (testing "Clerk fragments and comments render values from def vars"
+    (is (= 3
+           (-> (eval-test/eval+extract-doc-blocks "(ns nextjournal.clerk.viewer-test.fragments (:require [nextjournal.clerk :as clerk]))
+(clerk/fragment
+ 1 2 (def x 3))")
+               last :nextjournal/value :nextjournal/presented :nextjournal/value)))
+    (is (= 3
+           (-> (eval-test/eval+extract-doc-blocks "(ns nextjournal.clerk.viewer-test.fragments (:require [nextjournal.clerk :as clerk]))
+(clerk/comment
+ 1 2 (def x 3))")
+               last :nextjournal/value :nextjournal/presented :nextjournal/value)))
+
+    (is (= 4
+           (-> (eval-test/eval+extract-doc-blocks "(ns nextjournal.clerk.viewer-test.fragments (:require [nextjournal.clerk :as clerk]))
+(clerk/comment
+ 1 2 (clerk/comment 3 (def x 4)))")
+               last :nextjournal/value :nextjournal/presented :nextjournal/value)))
+
+    (is (= 4
+           (-> (eval-test/eval+extract-doc-blocks "(ns nextjournal.clerk.viewer-test.fragments (:require [nextjournal.clerk :as clerk]))
+(clerk/fragment
+ 1 (clerk/comment 2 (clerk/fragment 3 (def x 4))))")
+               last :nextjournal/value :nextjournal/presented :nextjournal/value))))
+
   (testing "Fragments emit distinct results for all of their (nested) children"
     (is (= 6
            (count